+++ src/modules/edc-demo/components/empty-state/empty-state.component.ts
@@ -12,5 +12,5 @@ export class EmptyStateComponent {
   cls = true;
 
   @Input()
-  data = '';
+  errorMessage = '';
 }