+++ src/modules/edc-demo/components/loading-state/loading-state.component.ts
@@ -3,7 +3,6 @@ import {Component, HostBinding} from '@angular/core';
 @Component({
   selector: 'edc-demo-loading-state',
   templateUrl: './loading-state.component.html',
-  styleUrls: ['./loading-state.component.scss'],
 })
 export class LoadingStateComponent {
   @HostBinding('class.flex')